Lower border with the program name.

+++ src/render.c
@@ -35,7 +35,7 @@ void clearWindow(void) { // To provide rendering in a one frame.
 	}
 }
 
-void window(char pressedKey, uint32_t charBuffer) {
+void window(char pressedKey, int8_t charBuffer) {
 	printf("%c", pressedKey); // Printing all chars will be here.
 	for(i = 2; i < windowSize('y'); i++) {
 		printf("%c", '\n');
